[Bronze II] Title: 숫자, Time: 192 ms, Memory: 39460 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/10093. 숫자/숫자.java
@@ -6,9 +6,9 @@ public class Main {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
         StringBuilder sb = new StringBuilder();
         StringTokenizer str = new StringTokenizer(br.readLine());
-        double A = Double.parseDouble(str.nextToken());
-        double B = Double.parseDouble(str.nextToken());
-        double temp;
+        long A = Long.parseLong(str.nextToken());
+        long B = Long.parseLong(str.nextToken());
+        long temp;
         if(A > B) {
             temp = A;
             A = B;
@@ -18,10 +18,10 @@ public class Main {
             sb.append(0 + "\n");
         }
         else {
-            sb.append((int)(B-A-1) + "\n");
+            sb.append((long)(B-A-1) + "\n");
         }
         for(double i= A+1 ; i <B; i++) {
-            sb.append((int)i + " ");
+            sb.append((long)i + " ");
         }
         sb.deleteCharAt(sb.length()-1);
         System.out.println(sb);
